rename some serialization files
