+++ CedarJava/build.gradle
@@ -76,7 +76,7 @@ configurations {
 }
 
 dependencies {
-    // Do not upgrade to Jackson 3.x without addressing stack overflow issues in ValueCedarDeserializer
+    // Do not upgrade to Jackson 3.x without addressing stack overflow issues in ValueDeserializer
     // The upgrade should be reviewed by AppSec
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.1'
     implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.17.1'
