+++ CedarJava/src/main/java/com/cedarpolicy/value/Value.java
@@ -16,11 +16,11 @@
 
 package com.cedarpolicy.value;
 
-import com.cedarpolicy.serializer.ValueCedarDeserializer;
+import com.cedarpolicy.serializer.ValueDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 /** A value in the Cedar language model. */
-@JsonDeserialize(using = ValueCedarDeserializer.class)
+@JsonDeserialize(using = ValueDeserializer.class)
 public abstract class Value {
     /**
      * Convert the Value instance into a string containing the Cedar source code for the equivalent
